requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
Django==5.0.3
django-cors-headers==4.3.1
django-filter==24.1
django-rest-knox==4.2.0
djangorestframework==3.14.0
dynaconf==3.2.4
hvac==2.1.0
idna==3.6
knox==0.1.14
loguru==0.7.2
Markdown==3.5.2
pillow==10.2.0
pycparser==2.21
PyJWT==2.8.0
pytz==2024.1
requests==2.31.0
shortuuid==1.0.13
six==1.16.0
sqlparse==0.4.4
urllib3==2.2.1
validators==0.22.0
